[Relay] Change imports to work as per normal usage.
