+++ types/react-relay/react-relay-tests.tsx
@@ -1,30 +1,383 @@
-import * as React from "react"
-import * as Relay from "react-relay"
+import * as React from "react";
+import {
+    Environment,
+    Network,
+    RecordSource,
+    Store,
+    ConnectionHandler,
+} from 'relay-runtime';
 
-interface Props {
-    text: string
-    userId: string
+////////////////////////////
+//  RELAY MODERN TESTS
+///////////////////////////
+import {
+    graphql,
+    commitMutation,
+    createFragmentContainer,
+    createPaginationContainer,
+    createRefetchContainer,
+    requestSubscription,
+    QueryRenderer,
+    ModernTypes,
+} from "react-relay";
+
+// ~~~~~~~~~~~~~~~~~~~~~
+// Modern Environment
+// ~~~~~~~~~~~~~~~~~~~~~
+function fetchQuery(operation: any, variables: any, cacheConfig: {}) {
+    return fetch('/graphql');
 }
+const network = Network.create(fetchQuery);
+const source = new RecordSource();
+const store = new Store(source);
+const modernEnvironment = new Environment({ network, store });
 
-interface Response {
+// ~~~~~~~~~~~~~~~~~~~~~
+// Modern QueryRenderer
+// ~~~~~~~~~~~~~~~~~~~~~
+const MyQueryRenderer = (props: { name: string}) => (
+  <QueryRenderer
+    environment={ modernEnvironment }
+    query={graphql`
+      query ExampleQuery($pageID: ID!) {
+        page(id: $pageID) {
+          name
+        }
+      }
+    `}
+    variables={{
+      pageID: '110798995619330',
+    }}
+    render={({error, props}) => {
+      if (error) {
+        return <div>{error.message}</div>;
+      } else if (props) {
+        return <div>{props.name} is great!</div>;
+      }
+      return <div>Loading</div>;
+    }}
+  />
+);
+
+// ~~~~~~~~~~~~~~~~~~~~~
+// Modern FragmentContainer
+// ~~~~~~~~~~~~~~~~~~~~~
+const MyFragmentContainer = createFragmentContainer(
+    class TodoListView extends React.Component {
+        render() {
+            return <div/>;
+        }
+    },
+    {
+        item: graphql`
+            fragment TodoItem_item on Todo {
+            text
+            isComplete
+            }
+        `,
+    }
+);
+
+// ~~~~~~~~~~~~~~~~~~~~~
+// Modern RefetchContainer
+// ~~~~~~~~~~~~~~~~~~~~~
+interface StoryInterface { id: string; }
+interface FeedStoriesProps {
+    relay: ModernTypes.RelayRefetchProp;
+    feed: {
+        stories: { edges: Array<{ node: StoryInterface }> }
+    };
 }
-
-export default class AddTweetMutation extends Relay.Mutation<Props, Response> {
-
-    getMutation () {
-        return Relay.QL`mutation{addTweet}`
+class Story extends React.Component<{ story: StoryInterface }> {}
+class FeedStories extends React.Component<FeedStoriesProps> {
+    render() {
+      return (
+        <div>
+          {this.props.feed.stories.edges.map(
+            edge => <Story story={edge.node} key={edge.node.id} />
+          )}
+          <button
+            onClick={() => this._loadMore}
+            title="Load More"
+          />
+        </div>
+      );
     }
 
-    getFatQuery () {
+    _loadMore() {
+      // Increments the number of stories being rendered by 10.
+      const refetchVariables = (fragmentVariables: {count: number }) => ({
+        count: fragmentVariables.count + 10,
+      });
+      this.props.relay.refetch(refetchVariables);
+    }
+  }
+
+const FeedRefetchContainer = createRefetchContainer(
+    FeedStories,
+    {
+      feed: graphql.experimental`
+        fragment FeedStories_feed on Feed
+        @argumentDefinitions(
+          count: {type: "Int", defaultValue: 10}
+        ) {
+          stories(first: $count) {
+            edges {
+              node {
+                id
+                ...Story_story
+              }
+            }
+          }
+        }
+      `
+    },
+    graphql.experimental`
+      query FeedStoriesRefetchQuery($count: Int) {
+        feed {
+          ...FeedStories_feed @arguments(count: $count)
+        }
+      }
+    `,
+  );
+
+// ~~~~~~~~~~~~~~~~~~~~~
+// Modern PaginationContainer
+// ~~~~~~~~~~~~~~~~~~~~~
+interface FeedProps {
+    user: { feed: { edges: Array<{ node: StoryInterface}>}};
+    relay: ModernTypes.RelayPaginationProp;
+}
+class Feed extends React.Component<FeedProps> {
+    render() {
+        return (<div>
+            {this.props.user.feed.edges.map(
+            edge => <Story story={edge.node} key={edge.node.id} />
+            )}
+            <button
+                onClick={() => this._loadMore()}
+                title="Load More"
+            />
+        </div>);
+    }
+
+    _loadMore() {
+        if (!this.props.relay.hasMore() || this.props.relay.isLoading()) {
+            return;
+        }
+
+        this.props.relay.loadMore(
+            10, // Fetch the next 10 feed items
+            e => {
+                console.log(e);
+            },
+        );
+    }
+}
+
+const FeedPaginationContainer = createPaginationContainer(
+    Feed,
+    {
+        user: graphql`
+        fragment Feed_user on User {
+            feed(
+            first: $count
+            after: $cursor
+            orderby: $orderBy # other variables
+            ) @connection(key: "Feed_feed") {
+            edges {
+                node {
+                id
+                ...Story_story
+                }
+            }
+            }
+        }
+        `,
+    },
+    {
+        direction: 'forward',
+        getConnectionFromProps(props: { user: {feed: any}}) {
+            return props.user && props.user.feed;
+        },
+        getFragmentVariables(prevVars, totalCount) {
+            return {
+                ...prevVars,
+                count: totalCount,
+            };
+        },
+        getVariables(props, {count, cursor}, fragmentVariables) {
+            return {
+                count,
+                cursor,
+                // in most cases, for variables other than connection filters like
+                // `first`, `after`, etc. you may want to use the previous values.
+                orderBy: fragmentVariables.orderBy,
+            };
+        },
+        query: graphql`
+        query FeedPaginationQuery(
+            $count: Int!
+            $cursor: String
+            $orderby: String!
+        ) {
+            user {
+            # You could reference the fragment defined previously.
+            ...Feed_user
+            }
+        }
+        `
+    }
+);
+
+// ~~~~~~~~~~~~~~~~~~~~~
+// Modern Mutations
+// ~~~~~~~~~~~~~~~~~~~~~
+const mutation = graphql`
+mutation MarkReadNotificationMutation(
+    $input: MarkReadNotificationData!
+) {
+    markReadNotification(data: $input) {
+        notification {
+            seenState
+        }
+    }
+}
+`;
+
+const optimisticResponse = {
+    markReadNotification: {
+        notification: {
+            seenState: 'SEEN',
+        },
+    },
+};
+
+const configs = [{
+    type: 'NODE_DELETE' as 'NODE_DELETE',
+    deletedIDFieldName: 'destroyedShipId',
+}, {
+    type: 'RANGE_ADD' as 'RANGE_ADD',
+    parentID: 'shipId',
+    connectionInfo: [{
+        key: 'AddShip_ships',
+        rangeBehavior: 'append',
+    }],
+    edgeName: 'newShipEdge',
+}, {
+    type: 'RANGE_DELETE' as 'RANGE_DELETE',
+    parentID: 'todoId',
+    connectionKeys: [{
+        key: 'RemoveTags_tags',
+        rangeBehavior: 'append',
+    }],
+    pathToConnection: ['todo', 'tags'],
+    deletedIDFieldName: 'removedTagId'
+}];
+
+function markNotificationAsRead(source: string, storyID: string) {
+    const variables = {
+        input: {
+            source,
+            storyID,
+        },
+    };
+
+    commitMutation(
+        modernEnvironment,
+        {
+            configs,
+            mutation,
+            optimisticResponse,
+            variables,
+            onCompleted: (response, errors) => {
+                console.log('Response received from server.');
+            },
+            onError: err => console.error(err),
+        },
+    );
+}
+
+// ~~~~~~~~~~~~~~~~~~~~~
+// Modern Subscriptions
+// ~~~~~~~~~~~~~~~~~~~~~
+const subscription = graphql`
+    subscription MarkReadNotificationSubscription(
+        $storyID: ID!
+    ) {
+    markReadNotification(storyID: $storyID) {
+        notification {
+        seenState
+        }
+    }
+    }
+`;
+const variables = {
+    storyID: '123',
+};
+requestSubscription(
+    modernEnvironment, // see Environment docs
+    {
+        subscription,
+        variables,
+        // optional but recommended:
+        onCompleted: () => {},
+        onError: error => console.error(error),
+        // example of a custom updater
+        updater: store => {
+            // Get the notification
+            const rootField = store.getRootField('markReadNotification');
+            const notification = !!rootField && rootField.getLinkedRecord('notification');
+            // Add it to a connection
+            const viewer = store.getRoot().getLinkedRecord('viewer');
+            const notifications =
+                ConnectionHandler.getConnection(viewer, 'notifications');
+            const edge = ConnectionHandler.createEdge(
+                store,
+                notifications,
+                notification,
+                '<TypeOfNotificationsEdge>',
+            );
+            ConnectionHandler.insertEdgeAfter(notifications, edge);
+        },
+    }
+);
+
+////////////////////////////
+//  RELAY COMPAT TESTS
+///////////////////////////
+import {
+    QueryRenderer as CompatQueryRenderer,
+    createFragmentContainer as createFragmentContainerCompat
+} from "react-relay/compat";
+
+// TODO? This is all more or less identical to modern...
+
+////////////////////////////
+//  RELAY-CLASSIC TESTS
+///////////////////////////
+import * as Relay from "react-relay/classic";
+
+interface Props {
+    text: string;
+    userId: string;
+}
+
+export default class AddTweetMutation extends Relay.Mutation<Props, {}> {
+    getMutation() {
+        return Relay.QL`mutation{addTweet}`;
+    }
+
+    getFatQuery() {
         return Relay.QL`
             fragment on AddTweetPayload {
                 tweetEdge
                 user
             }
-        `
+        `;
     }
 
-    getConfigs () {
+    getConfigs() {
         return [{
             type: "RANGE_ADD",
             parentName: "user",
@@ -34,19 +387,19 @@ export default class AddTweetMutation extends Relay.Mutation<Props, Response> {
             rangeBehaviors: {
                 "": "append",
             },
-        }]
+        }];
     }
 
-    getVariables () {
-        return this.props
+    getVariables() {
+        return this.props;
     }
 }
 
 interface ArtworkProps {
     artwork: {
         title: string
-    },
-    relay: Relay.RelayProp,
+    };
+    relay: Relay.RelayProp;
 }
 
 class Artwork extends React.Component<ArtworkProps> {
@@ -55,7 +408,7 @@ class Artwork extends React.Component<ArtworkProps> {
             <a href={`/artworks/${this.props.relay.variables.artworkID}`}>
                 {this.props.artwork.title}
             </a>
-        )
+        );
     }
 }
 
@@ -67,7 +420,7 @@ const ArtworkContainer = Relay.createContainer(Artwork, {
             }
         `
     }
-})
+});
 
 class StubbedArtwork extends React.Component {
     render() {
@@ -83,10 +436,10 @@ class StubbedArtwork extends React.Component {
                 setVariables: () => {},
                 forceFetch: () => {},
                 hasOptimisticUpdate: () => false,
-                getPendingTransactions: (): Relay.RelayMutationTransaction[] => undefined,
+                getPendingTransactions: (): any => undefined,
                 commitUpdate: () => {},
             }
-        }
-        return <ArtworkContainer {...props} />
+        };
+        return <ArtworkContainer {...props} />;
     }
 }
