--- types/react-relay/v1/tsconfig.json
@@ -1,33 +1,0 @@
-{
-    "compilerOptions": {
-        "module": "commonjs",
-        "lib": [
-            "es6",
-            "dom"
-        ],
-        "noImplicitAny": true,
-        "noImplicitThis": true,
-        "strictNullChecks": true,
-        "baseUrl": "../../",
-        "typeRoots": [
-            "../../"
-        ],
-        "types": [],
-        "noEmit": true,
-        "forceConsistentCasingInFileNames": true,
-        "jsx": "react",
-        "paths": {
-            "react-relay": ["react-relay/v1"],
-            "react-relay/*": ["react-relay/v1/*"],
-            "react-relay/classic": ["react-relay/v1/classic.d"],
-            "react-relay/compat": ["react-relay/v1/compat.d"]
-        }
-    },
-    "files": [
-        "index.d.ts",
-        "react-relay-tests.tsx",
-        "lib/react-relay-classic.d.ts",
-        "lib/react-relay-modern.d.ts",
-        "lib/react-relay-compat.d.ts"
-    ]
-}
